Refactor sprite positioning for accurate raycasting
Changes were made to improve sprite positioning for more accurate raycasting results. In 'update_player_bonus.c', 'hit_sprite_bonus.c', and 'fill_sprites_buffers_bonus.c', coordinates were revised for correct mapping. Now, they accurately reflect the sprites' positions in the grid, improving collision detection and rendering processes. These changes enhance the game's overall performance and visual experience.

+++ bonus/srcs/sprites/hit_sprite_bonus.c
@@ -6,7 +6,7 @@
 /*   By: beroux <beroux@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/01 00:52:06 by beroux            #+#    #+#             */
-/*   Updated: 2023/10/05 06:14:57 by beroux           ###   ########.fr       */
+/*   Updated: 2023/10/08 18:29:29 by beroux           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ static t_sprites_list	*new_node(int pos[2], t_sprite *sprite, \
 									double player_pos[2])
 {
 	t_sprites_list	*node;
-	int				true_pos[2];
+	double			true_pos[2];
 
 	node = malloc(sizeof (t_sprites_list));
 	if (!node)
